fix: address all 5 code review findings from Batch 1
1. (High) Alembic bootstrap: env.py now reads DATUM_DATABASE_URL from
   environment; hand-crafted migration 001 checked in with all 8 tables
   in dependency order including proper downgrade.

2. (Medium) Path traversal: validate_canonical_path() rejects absolute
   paths and parent traversal (../) before doc_manifest_dir resolves.
   8 new tests cover valid paths, traversal, and absolute rejection.

3. (Medium) Compose self-consistency: watcher.py stub module added so
   `python -m datum.watcher` doesn't fail with ModuleNotFoundError.
   Real implementation deferred to Task 8.

4. (Medium) Schema type drift: byte_size columns on source_files and
   document_versions changed from Integer to BigInteger to match design
   doc BIGINT specification. Migration uses sa.BigInteger.

5. (Medium) Test depth: added test_alembic_migration_revision_exists,
   test_byte_size_columns_are_bigint, test_alembic_env_reads_database_url,
   and 8 path validation tests. Total: 24 tests, all passing.

+++ backend/datum/models/core.py
@@ -2,7 +2,7 @@ from datetime import datetime
 from typing import Optional
 from uuid import UUID
 
-from sqlalchemy import ARRAY, DateTime, ForeignKey, Integer, String, Text, UniqueConstraint
+from sqlalchemy import ARRAY, BigInteger, DateTime, ForeignKey, Integer, String, Text, UniqueConstraint
 from sqlalchemy.dialects.postgresql import JSONB, UUID as PG_UUID
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
@@ -64,7 +64,7 @@ class SourceFile(Base):
     canonical_path: Mapped[str] = mapped_column(String, nullable=False)
     object_kind: Mapped[str] = mapped_column(String, nullable=False)
     content_hash: Mapped[str] = mapped_column(String, nullable=False)
-    byte_size: Mapped[Optional[int]] = mapped_column(Integer)
+    byte_size: Mapped[Optional[int]] = mapped_column(BigInteger)
     mtime: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True))
     last_seen_at: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True))
     indexed_at: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True))
@@ -113,7 +113,7 @@ class DocumentVersion(Base):
     content_hash: Mapped[str] = mapped_column(String, nullable=False)
     filesystem_path: Mapped[str] = mapped_column(String, nullable=False)
     content_type: Mapped[Optional[str]] = mapped_column(String)
-    byte_size: Mapped[Optional[int]] = mapped_column(Integer)
+    byte_size: Mapped[Optional[int]] = mapped_column(BigInteger)
     label: Mapped[Optional[str]] = mapped_column(String)
     change_source: Mapped[Optional[str]] = mapped_column(String)
     agent_name: Mapped[Optional[str]] = mapped_column(String)
